+++ db/migrate/20200512144535_add_ingested_boolean_to_transactions.rb
@@ -1,0 +1,5 @@
+class AddIngestedBooleanToTransactions < ActiveRecord::Migration[6.0]
+  def change
+    add_column :transactions, :ingested, :boolean, default: false
+  end
+end